requirements.txt
Django==3.2
Pillow==8.2.0
pymorphy2==0.9.1
pandas==1.4.2
scikit-learn==1.0.2
django-crispy-forms==1.12.0
django-notifications-hq==1.6.0
APScheduler==3.6.3
requests==2.26.0
bs4==0.0.1
beautifulsoup4==4.10.0